backend/pyproject.toml
[tool.poetry]
name = "initry"
version = "0.1.0"
description = ""
authors = ["Andrejs Smirnovs"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.110.1"
pymongo = "4.6.3"
pydantic = "2.6.4"
uvicorn = "0.29.0"
celery = "5.3.6"
redis = "5.0.3"
python-dotenv = "1.0.1"
websockets = "^12.0"
grpcio = "^1.62.1"
grpcio-tools = "^1.62.1"

[tool.poetry.dev-dependencies]
flower = "2.0.1"
black = "24.3.0"
pylint = "3.1.0"
flake8 = "7.0.0"
isort = "5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
